Move the backport of sympy.Printable into its own file
Since this is a temporary solution until we drop sympy 1.6, putting it in its own file keeps it isolated.
